Move all plugin logic into knife supermarket
We want to deprecate knife cookbook site. This moves all the logic into
the knife supermarket plugins and has the legacy knife cookbook site
plugins inherit from that.
