+++ lib/chef/knife/cookbook_site_unshare.rb
@@ -18,45 +18,18 @@
 #
 
 require "chef/knife"
+require "chef/knife/supermarket_unshare"
 
 class Chef
   class Knife
-    class CookbookSiteUnshare < Knife
+    class CookbookSiteUnshare < Knife::SupermarketUnshare
 
-      deps do
-        require "chef/json_compat"
-      end
+    # Handle the subclassing (knife doesn't do this :()
+      dependency_loaders.concat(superclass.dependency_loaders)
+      options.merge!(superclass.options)
 
-      banner "knife cookbook site unshare COOKBOOK"
+      banner "knife cookbook site unshare COOKBOOK (options)"
       category "cookbook site"
-
-      option :supermarket_site,
-        short: "-m SUPERMARKET_SITE",
-        long: "--supermarket-site SUPERMARKET_SITE",
-        description: "Supermarket Site",
-        default: "https://supermarket.chef.io",
-        proc: Proc.new { |supermarket| Chef::Config[:knife][:supermarket_site] = supermarket }
-
-      def run
-        @cookbook_name = @name_args[0]
-        if @cookbook_name.nil?
-          show_usage
-          ui.fatal "You must provide the name of the cookbook to unshare"
-          exit 1
-        end
-
-        confirm "Do you really want to unshare all versions of the cookbook #{@cookbook_name}"
-
-        begin
-          rest.delete "#{config[:supermarket_site]}/api/v1/cookbooks/#{@name_args[0]}"
-        rescue Net::HTTPServerException => e
-          raise e unless e.message =~ /Forbidden/
-          ui.error "Forbidden: You must be the maintainer of #{@cookbook_name} to unshare it."
-          exit 1
-        end
-
-        ui.info "Unshared all versions of the cookbook #{@cookbook_name}"
-      end
 
     end
   end
